Move everything to one model file and add unit test
