+++ torchtitan/models/llama_multimodal/__init__.py
@@ -1,0 +1,16 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+#
+# Llama 2 is licensed under the LLAMA 2 Community License,
+# Copyright (c) Meta Platforms, Inc. All Rights Reserved.
+
+from torchtitan.models.llama_multimodal.model import ModelArgs, VisionEncoder
+
+__all__ = ["VisionEncoder", "ModelArgs"]
+
+llama3_2_configs = {
+    # TODO: add configs for llama3.2
+}
