--- torchtitan/models/llama_multimodal/norms.py
@@ -1,38 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-#
-# Llama 2 is licensed under the LLAMA 2 Community License,
-# Copyright (c) Meta Platforms, Inc. All Rights Reserved.
-
-from typing import Any
-
-import torch
-import torch.nn as nn
-
-
-class Fp32LayerNorm(nn.LayerNorm):
-    """
-    Wrapper around :class:`~torch.nn.LayerNorm` to support mixed-precision training.
-    """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """
-        Args:
-            x (torch.Tensor): Input tensor.
-        Returns:
-            torch.Tensor: The normalized output tensor having the same shape as ``x``.
-        """
-        output = nn.functional.layer_norm(
-            x.float(),
-            self.normalized_shape,
-            self.weight.float() if self.weight is not None else None,
-            self.bias.float() if self.bias is not None else None,
-            self.eps,
-        )
-        return output.type_as(x)
